feat(targets): add version revision tracking
to display the configuration version of the system you can use the
following command in the `ghaf-host`.

`nixos-version --json`

in addition the generation label will also be changed to reflect the
version. This can be seen during the boot.

If there are uncommitted changes there will be a `-dirty` moniker
appended to the hash. The hash directly matches the git commit from
which the build was made.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>